requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.0.2
SQLAlchemy>=2.0.36
psycopg2-binary==2.9.10

# Security
Flask-Bcrypt==1.0.1
bcrypt==4.0.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# Utilities
python-dotenv==1.0.0
pdfkit==1.0.0
markupsafe==2.1.3

# Template engine (comes with Flask but pinned for consistency)
Jinja2==3.1.2
itsdangerous==2.1.2

# Other dependencies
blinker==1.7.0
click==8.1.7
colorama==0.4.6
